requirements.txt
certifi==2018.8.24
chardet==3.0.4
Django==1.11.15
django-filter==2.0.0
djangorestframework==3.8.2
idna==2.7
Markdown==2.6.11
MySQL-python==1.2.5
pytz==2018.5
requests==2.19.1
urllib3==1.23
